+++ systole/detectors/__init__.py
@@ -1,5 +1,7 @@
+from .hamilton import hamilton
 from .pan_tompkins import pan_tompkins
 
 __all__ = [
     "pan_tompkins",
+    "hamilton",
 ]